song-agent/pyproject.toml
[project]
name = "song-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckduckgo-search>=7.5.0",
    "google-genai>=1.3.0",
    "ipython>=8.0.0",
    "python-dotenv>=1.0.1",
]
